test: re-write stream test to work as a benchmark as well
